+++ examples/data-sources/iosxr_error_disable_recovery/data-source.tf
@@ -1,0 +1,2 @@
+data "iosxr_error_disable_recovery" "example" {
+}