+++ examples/data-sources/iosxr_tag_set/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_tag_set" "example" {
+  set_name = "TEST"
+}